extract header user and invite button
